+++ petshop/forms.py
@@ -1,14 +1,14 @@
 import babel
-from wtforms.validators import DataRequired, Email
 from flask_wtf import FlaskForm
-from wtforms import (
-                    StringField, SubmitField, SelectField, IntegerField,
-                    DecimalField, DateField, SelectMultipleField
+from datetime import datetime
+from wtforms_components import (
+                    StringField, SelectField, IntegerField,
+                    DecimalField, DateField,
+                    SelectMultipleField, DateRange
                     )
+from wtforms.fields import SubmitField
+from wtforms.validators import DataRequired, Email
 # import dash_core_components as dcc
-# from datetime import datetime as dt
-
-# from wtforms.fields.html5 import DateField
 
 
 class Form_clientes(FlaskForm):
@@ -20,16 +20,16 @@ class Form_clientes(FlaskForm):
                         validators=[DataRequired()]
                         )
 
-    tel1 = StringField('tel1', validators=[DataRequired()])
-    tel2 = StringField('tel2', validators=[DataRequired()])
-    email = StringField('e-mail', validators=[DataRequired(), Email()])
+    tel1 = StringField('Tel.1', validators=[DataRequired()])
+    tel2 = StringField('Tel.2', validators=[DataRequired()])
+    email = StringField('E-mail', validators=[DataRequired(), Email()])
 
     rua = StringField('Rua', validators=[DataRequired()])
-    numero = IntegerField('Numero', validators=[DataRequired()])
+    numero = IntegerField('Número', validators=[DataRequired()])
     bairro = StringField('Bairo', validators=[DataRequired()])
-    cidade = StringField('Cidade', validators=[DataRequired()])
-    estado = StringField('Estado', validators=[DataRequired()])
-    distancia = DecimalField('Distancia', validators=[DataRequired()])
+    cidade = StringField('Cidade', default='Santo André', validators=[DataRequired()])
+    estado = StringField('Estado', default='SP', validators=[DataRequired()])
+    distancia = DecimalField('Distância', default=5, validators=[DataRequired()])
 
     submit = SubmitField('Adiciona cliente')
 
@@ -44,7 +44,7 @@ class Form_peludos(FlaskForm):
                         'Sexo', choices=[('M', 'Macho'), ('F', 'Fêmea')],
                         validators=[DataRequired()]
                         )
-    breed = StringField('Raça', validators=[DataRequired()])
+    breed = StringField('Raça', default='Shitzu', validators=[DataRequired()])
     pelagem = SelectField(
                         'Tipo de pelo',
                         choices=[('curto', 'Curto'), ('longo', 'Longo')],
@@ -52,14 +52,14 @@ class Form_peludos(FlaskForm):
                         )
 
     nascimento = DateField(
-                            'Data de Nascimento', format='%d-%m-%Y',
+                            'Data de Nascimento',
                             validators=[DataRequired()]
                             )
 
-    data_start = DateField('Date de início no pet', format='%d-%m-%Y',
+    data_start = DateField('Date de início no pet', # format='%d-%m-%Y',
                            validators=[DataRequired()])
     castrado = SelectField(
-                            'Peludo foi castrado?',
+                            'É castrado?',
                             choices=[('S', 'Sim'), ('N', 'Não')],
                             validators=[DataRequired()]
                             )
@@ -77,17 +77,16 @@ class Form_vendas_intro(FlaskForm):
 class Form_vendas(FlaskForm):
     """Form venda."""
 
+    # cliente = SelectField(u'Dono do peludo', choices=[], coerce=int)
+
     peludos = SelectMultipleField('Peludo(s)', choices=[], coerce=int)
     descricao = StringField('Descrição', validators=[DataRequired()])
-    data_venda = DateField(
-        'Data da venda', format='%d-%m-%Y', validators=[DataRequired()])
 
-
-    valor_venda = DecimalField('Valor da venda',  # use_locale=True,
+    valor_servicos = DecimalField('Valor de serviços',  # use_locale=True,
                                validators=[DataRequired()])
-    valor_taxi = DecimalField('Valor do taxi dog',  # use_locale=True,
+    valor_taxi = DecimalField('Valor do taxi dog', default=0,  # use_locale=True,
                               validators=[DataRequired()])
-    n_banhos = IntegerField('Quantidade de banhos',
+    n_banhos = IntegerField('Quantidade de banhos', default=1,
                             validators=[DataRequired()])
     tipo = SelectField('Avulso ou pacote',
                        choices=[
@@ -97,7 +96,9 @@ class Form_vendas(FlaskForm):
                        validators=[DataRequired()]
                        )
 
-    pacote = StringField('Tipo de pacote', validators=[DataRequired()])
+    pacote = StringField('Tipo de pacote', default="puka",
+                         validators=[DataRequired()])
+
     forma_pagto = SelectField('Forma de pagamento',
                               choices=[
                                         ('cash', 'cash'),
@@ -109,8 +110,9 @@ class Form_vendas(FlaskForm):
                               validators=[DataRequired()]
                               )
 
-    data_pagto = DateField('Data de pagamento',
-                           format='%d-%m-%Y', validators=[DataRequired()])
+    data_pagto = DateField('Data de pagamento', default=datetime.today(),
+                          validators=[DataRequired()])
+
     valor_entrada = DecimalField(
         'Valor real de entrada',  # use_locale=True,
         validators=[DataRequired()])