+++ changelog.md
@@ -1,3 +1,10 @@
-# Changelog
+# Changelog | todo
+
+## 2018-11-02
+
+- better looking forms - html5
+- api
+- relatório vendas shows after cadastro_vendas
+
 
 - 2018-10-20: testing models